Puntos

+++ backend/Ticketcine/src/main/java/com/s1511/ticketcine/application/dto/login/ResponseLogin.java
@@ -6,6 +6,6 @@ public record ResponseLogin(
         String id,
         String firstName,
         String lastName,
-        Long moviePoints
+        Double moviePoints
 ) {
 }
